workspace-db totally gone
