+++ src/core/main.ts
@@ -1,12 +1,13 @@
 import { __entry } from './entry'
+import Callbag from 'src/utilities/callbag'
+import { codeBlockCreated } from 'src/new/codeBlockHandler'
+import { iife } from 'src/utilities/utilities'
+
+
 export const plugin = __entry.plugin
-
-import { createDb } from 'src/workspace/db-schema'
-
 export const svgs = new Map()
-export const workspace = createDb()
 
-~(async () => {
+iife(async () => {
   import('./events')
   import('src/internal-links/handle-internal-links')
   await import('src/new/file-settings-button')
@@ -15,9 +16,10 @@ export const workspace = createDb()
   const { loadStyleFeatures } = await import('src/rendering/style-features')
   const { GlobalSettingsDialog } = await import('src/settings/dialogs')
   const { codeBlockHandler } = await import('src/new/codeBlockHandler')
+  const { CodeBlockRenderer } = await import('src/rendering/renderer-codeblock')
+  Callbag.subscribe(codeBlockCreated, CodeBlockRenderer)
 
   plugin.addSettingTab(new GlobalSettingsDialog())
   plugin.registerMarkdownCodeBlockProcessor('markmap', codeBlockHandler)
   loadStyleFeatures()
-  import('src/workspace')
-})()
+})
